separated project removeResources modal and actions
